refactor(starter/nuxt-auth): extract auth content into SFC
