+++ starter/nuxt-auth/components/auth/AuthShell.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="md:h-screen grid grid-cols-1 md:grid-cols-2 gap-4">
+    <slot />
+  </div>
+</template>
