+++ starter/nuxt-auth/components/auth/AuthSidebar.vue
@@ -1,0 +1,5 @@
+<template>
+  <div
+    class="bg-gradient-to-tr from-primary-500 via-indigo-500 to-cyan-500"
+  ></div>
+</template>
